refactor(chatRoom): 채팅 룸 컴포넌트로 코드 정리
