+++ src/views/layouts/NavigationBar.tsx
@@ -48,12 +48,11 @@ const NavigationBarStyle = styled.div`
 
   .navbar {
     display: flex;
-    position: fixed;
     flex-direction: column;
     justify-content: space-between;
     width: 6%;
     min-width: 80px;
-    height: 100vh;
+    height: 100%;
     background: black;
     padding: 30px 0;
   }
