+++ src/views/pages/ChatLists.tsx
@@ -1,11 +1,10 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import { MainContainer } from '../../components/styles/Home';
 function ChatLists() {
   const location = useLocation();
-
+  useEffect(() => {});
   const { data } = location.state;
-  console.log(data);
   return <MainContainer>{data}</MainContainer>;
 }
